Button PR Changes
Renamed button name for consistency with other app buttons.

Added in global app color references since we already have defined colors.

The button was not actually using it's props, so tweaked that.

Ran ESLint to fix formatting issues.

Changed the button's function header to reduce the number of lines of code.